@@ -1,5 +1,6 @@
 package com.foxit.flutterfoxitpdf;
 
+import androidx.annotation.NonNull;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,42 +8,69 @@ import android.os.Bundle;
 import com.foxit.sdk.common.Constants;
 import com.foxit.sdk.common.Library;
 
+import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.embedding.engine.plugins.activity.ActivityAware;
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
-import io.flutter.plugin.common.PluginRegistry.Registrar;
 
-/** FlutterFoxitpdfPlugin */
-public class FlutterFoxitpdfPlugin implements MethodCallHandler {
-  /** Plugin registration. */
-  public static void registerWith(Registrar registrar) {
-    final MethodChannel channel = new MethodChannel(registrar.messenger(), "flutter_foxitpdf");
-    channel.setMethodCallHandler(new FlutterFoxitpdfPlugin(registrar.activity()));
+/** FlutterfoxitpdfPlugin */
+public class FlutterfoxitpdfPlugin implements FlutterPlugin, MethodCallHandler, ActivityAware {
+  /// The MethodChannel that will the communication between Flutter and native Android
+  ///
+  /// This local reference serves to register the plugin with the Flutter Engine and unregister it
+  /// when the Flutter Engine is detached from the Activity
+  // private ActivityPluginBinding _activityBinding;
+  private Activity mActivity;
+  private MethodChannel channel;
+  private int errorCode = Constants.e_ErrUnknown;
+
+
+  @Override
+  public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
+    channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "flutterfoxitpdf");
+    channel.setMethodCallHandler(this);
+
+  }
+  @Override
+  public void onAttachedToActivity(ActivityPluginBinding binding) {
+    mActivity = binding.getActivity();
   }
 
-  private int errorCode = Constants.e_ErrUnknown;
-  private Activity mActivity = null;
-  private FlutterFoxitpdfPlugin(Activity activity) {
-    mActivity = activity;
+
+  @Override
+  public void   onDetachedFromActivity() {
+    mActivity = null;
   }
 
   @Override
-  public void onMethodCall(MethodCall call, Result result) {
-    switch (call.method) {
-      case "getPlatformVersion":
-          result.success("Android " + android.os.Build.VERSION.RELEASE);
-          break;
-      case "initialize":
-        initialize(call, result);
-        break;
-      case "openDocument":
-        openDocument(call, result);
-        break;
-      default:
-        result.notImplemented();
-        break;
+  public void   onReattachedToActivityForConfigChanges​(ActivityPluginBinding binding) {
+    mActivity =  binding.getActivity();
+  }
+
+  @Override
+  public void   onDetachedFromActivityForConfigChanges() {
+    mActivity = null;
+  }
+
+  @Override
+  public void onMethodCall(@NonNull MethodCall call, @NonNull Result result) {
+    if (call.method.equals("getPlatformVersion")) {
+      result.success("Android " + android.os.Build.VERSION.RELEASE);
+    }  else if (call.method.equals("initialize")) {
+      initialize(call, result);
+    } else if (call.method.equals("openDocument")) {
+      openDocument(call, result);
+    } else {
+      result.notImplemented();
     }
+  }
+
+  @Override
+  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
+    channel.setMethodCallHandler(null);
   }
 
   private void initialize(MethodCall call, Result result) {
